Revert "Merge pull request #18972 from CyrusNajmabadi/docHighlights2"
This reverts commit 8262112e63d0c92bfa421bcb49e54b6a791cabbd, reversing
changes made to 837ca2141fc63d42ea4f6310c07e58624f0ca48a.
