+++ src/EditorFeatures/CSharp/ReferenceHighlighting/ReferenceHighlightingAdditionalReferenceProvider.cs
@@ -1,31 +1,32 @@
 // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System.Collections.Immutable;
+using System.Collections.Generic;
 using System.Composition;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.DocumentHighlighting;
+using Microsoft.CodeAnalysis.Editor.Implementation.ReferenceHighlighting;
 using Microsoft.CodeAnalysis.Host.Mef;
+using Roslyn.Utilities;
 
-namespace Microsoft.CodeAnalysis.CSharp.DocumentHighlighting
+namespace Microsoft.CodeAnalysis.Editor.CSharp.ReferenceHighlighting
 {
-    [ExportLanguageService(typeof(IDocumentHighlightsService), LanguageNames.CSharp), Shared]
-    internal class CSharpDocumentHighlightsService : AbstractDocumentHighlightsService
+    [ExportLanguageService(typeof(IReferenceHighlightingAdditionalReferenceProvider), LanguageNames.CSharp), Shared]
+    internal class ReferenceHighlightingAdditionalReferenceProvider : IReferenceHighlightingAdditionalReferenceProvider
     {
-        protected override async Task<ImmutableArray<Location>> GetAdditionalReferencesAsync(
+        public async Task<IEnumerable<Location>> GetAdditionalReferencesAsync(
             Document document, ISymbol symbol, CancellationToken cancellationToken)
         {
             // The FindRefs engine won't find references through 'var' for performance reasons.
             // Also, they are not needed for things like rename/sig change, and the normal find refs
-            // feature.  However, we would like the results to be highlighted to get a good experience
+            // feature.  However, we would lke the results to be highlighted to get a good experience
             // while editing (especially since highlighting may have been invoked off of 'var' in
             // the first place).
             //
             // So we look for the references through 'var' directly in this file and add them to the
             // results found by the engine.
-            var results = ArrayBuilder<Location>.GetInstance();
+            List<Location> results = null;
 
             if (symbol is INamedTypeSymbol && symbol.Name != "var")
             {
@@ -51,13 +52,18 @@ namespace Microsoft.CodeAnalysis.CSharp.DocumentHighlighting
 
                         if (originalSymbol.Equals(boundSymbol))
                         {
+                            if (results == null)
+                            {
+                                results = new List<Location>();
+                            }
+
                             results.Add(type.GetLocation());
                         }
                     }
                 }
             }
 
-            return results.ToImmutableAndFree();
+            return results ?? SpecializedCollections.EmptyEnumerable<Location>();
         }
     }
 }
