+++ src/EditorFeatures/Core/EditorFeatures.csproj
@@ -168,14 +168,6 @@
     <Compile Include="Implementation\Structure\BlockTagState.cs" />
     <Compile Include="Implementation\Suggestions\SuggestedActionSetComparer.cs" />
     <Compile Include="Implementation\Suggestions\SuggestedActionsSource.cs" />
-    <Compile Include="ReferenceHighlighting\NagivateToHighlightReferenceCommandHandler.cs" />
-    <Compile Include="ReferenceHighlighting\NagivateToHighlightReferenceCommandHandler.StartComparer.cs" />
-    <Compile Include="ReferenceHighlighting\ReferenceHighlightingViewTaggerProvider.cs" />
-    <Compile Include="ReferenceHighlighting\Tags\DefinitionHighlightTag.cs" />
-    <Compile Include="ReferenceHighlighting\Tags\DefinitionHighlightTagDefinition.cs" />
-    <Compile Include="ReferenceHighlighting\Tags\ReferenceHighlightTag.cs" />
-    <Compile Include="ReferenceHighlighting\Tags\WrittenReferenceHighlightTag.cs" />
-    <Compile Include="ReferenceHighlighting\Tags\WrittenReferenceHighlightTagDefinition.cs" />
     <Compile Include="SymbolSearch\IAddReferenceDatabaseWrapper.cs" />
     <Compile Include="SymbolSearch\IDatabaseFactoryService.cs" />
     <Compile Include="SymbolSearch\IDelayService.cs" />
@@ -638,6 +630,8 @@
     <Compile Include="Implementation\Peek\PeekableItemSource.cs" />
     <Compile Include="Implementation\Peek\PeekableItemSourceProvider.cs" />
     <Compile Include="Implementation\Peek\PeekHelpers.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\WrittenReferenceHighlightTag.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\WrittenReferenceHighlightTagDefinition.cs" />
     <Compile Include="Implementation\RenameTracking\IRenameTrackingLanguageHeuristicsService.cs" />
     <Compile Include="Implementation\RenameTracking\RenameTrackingCancellationCommandHandler.cs" />
     <Compile Include="Implementation\RenameTracking\RenameTrackingTagDefinition.cs" />
@@ -649,6 +643,15 @@
     <Compile Include="Implementation\Preview\PreviewConflictViewTaggerProvider.cs" />
     <Compile Include="Implementation\Preview\PreviewFactoryService.cs" />
     <Compile Include="Implementation\Preview\PreviewWarningViewTaggerProvider.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\AbstractDocumentHighlightsService.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\DefinitionHighlightTag.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\DefinitionHighlightTagDefinition.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\IDocumentHighlightsService.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\IReferenceHighlightingAdditionalReferenceProvider.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\NagivateToHighlightReferenceCommandHandler.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\NagivateToHighlightReferenceCommandHandler.StartComparer.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\ReferenceHighlightingViewTaggerProvider.cs" />
+    <Compile Include="Implementation\ReferenceHighlighting\ReferenceHighlightTag.cs" />
     <Compile Include="Implementation\RenameTracking\AbstractRenameTrackingCodeFixProvider.cs" />
     <Compile Include="Implementation\RenameTracking\RenameTrackingDiagnosticAnalyzer.cs" />
     <Compile Include="Implementation\RenameTracking\RenameTrackingTaggerProvider.cs" />
