+++ src/EditorFeatures/Core/Implementation/ReferenceHighlighting/AbstractDocumentHighlightsService.cs
@@ -14,46 +14,11 @@ using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
 
-namespace Microsoft.CodeAnalysis.DocumentHighlighting
+namespace Microsoft.CodeAnalysis.Editor.Implementation.ReferenceHighlighting
 {
-    internal abstract partial class AbstractDocumentHighlightsService : IDocumentHighlightsService
+    internal abstract class AbstractDocumentHighlightsService : IDocumentHighlightsService
     {
         public async Task<ImmutableArray<DocumentHighlights>> GetDocumentHighlightsAsync(
-            Document document, int position, IImmutableSet<Document> documentsToSearch, CancellationToken cancellationToken)
-        {
-            var (succeeded, highlights) = await GetDocumentHighlightsInRemoteProcessAsync(
-                document, position, documentsToSearch, cancellationToken).ConfigureAwait(false);
-
-            if (succeeded)
-            {
-                return highlights;
-            }
-
-            return await GetDocumentHighlightsInCurrentProcessAsync(
-                document, position, documentsToSearch, cancellationToken).ConfigureAwait(false);
-        }
-
-        private async Task<(bool succeeded, ImmutableArray<DocumentHighlights> highlights)> GetDocumentHighlightsInRemoteProcessAsync(
-            Document document, int position, IImmutableSet<Document> documentsToSearch, CancellationToken cancellationToken)
-        {
-            using (var session = await TryGetRemoteSessionAsync(
-                document.Project.Solution, cancellationToken).ConfigureAwait(false))
-            {
-                if (session == null)
-                {
-                    return (succeeded: false, ImmutableArray<DocumentHighlights>.Empty);
-                }
-
-                var result = await session.InvokeAsync<SerializableDocumentHighlights[]>(
-                    nameof(IRemoteDocumentHighlights.GetDocumentHighlightsAsync),
-                    document.Id,
-                    position,
-                    documentsToSearch.Select(d => d.Id).ToArray()).ConfigureAwait(false);
-                return (true, SerializableDocumentHighlights.Rehydrate(result, document.Project.Solution));
-            }
-        }
-
-        private async Task<ImmutableArray<DocumentHighlights>> GetDocumentHighlightsInCurrentProcessAsync(
             Document document, int position, IImmutableSet<Document> documentsToSearch, CancellationToken cancellationToken)
         {
             // use speculative semantic model to see whether we are on a symbol we can do HR
@@ -76,11 +41,11 @@ namespace Microsoft.CodeAnalysis.DocumentHighlighting
 
             // Get unique tags for referenced symbols
             return await GetTagsForReferencedSymbolAsync(
-                new SymbolAndProjectId(symbol, document.Project.Id), documentsToSearch,
+                new SymbolAndProjectId(symbol, document.Project.Id), documentsToSearch, 
                 solution, cancellationToken).ConfigureAwait(false);
         }
 
-        private static async Task<ISymbol> GetSymbolToSearchAsync(Document document, int position, SemanticModel semanticModel, ISymbol symbol, CancellationToken cancellationToken)
+        private async Task<ISymbol> GetSymbolToSearchAsync(Document document, int position, SemanticModel semanticModel, ISymbol symbol, CancellationToken cancellationToken)
         {
             // see whether we can use the symbol as it is
             var currentSemanticModel = await document.GetSemanticModelAsync(cancellationToken).ConfigureAwait(false);
@@ -117,7 +82,7 @@ namespace Microsoft.CodeAnalysis.DocumentHighlighting
             return ImmutableArray<DocumentHighlights>.Empty;
         }
 
-        private static bool ShouldConsiderSymbol(ISymbol symbol)
+        private bool ShouldConsiderSymbol(ISymbol symbol)
         {
             switch (symbol.Kind)
             {
@@ -167,13 +132,19 @@ namespace Microsoft.CodeAnalysis.DocumentHighlighting
                 documentsToSearch, cancellationToken).ConfigureAwait(false);
         }
 
-        protected virtual Task<ImmutableArray<Location>> GetAdditionalReferencesAsync(
+        private Task<IEnumerable<Location>> GetAdditionalReferencesAsync(
             Document document, ISymbol symbol, CancellationToken cancellationToken)
         {
-            return SpecializedTasks.EmptyImmutableArray<Location>();
+            var additionalReferenceProvider = document.Project.LanguageServices.GetService<IReferenceHighlightingAdditionalReferenceProvider>();
+            if (additionalReferenceProvider != null)
+            {
+                return additionalReferenceProvider.GetAdditionalReferencesAsync(document, symbol, cancellationToken);
+            }
+
+            return Task.FromResult(SpecializedCollections.EmptyEnumerable<Location>());
         }
 
-        private static async Task<ImmutableArray<DocumentHighlights>> CreateSpansAsync(
+        private async Task<ImmutableArray<DocumentHighlights>> CreateSpansAsync(
             Solution solution,
             ISymbol symbol,
             IEnumerable<ReferencedSymbol> references,
@@ -281,7 +252,7 @@ namespace Microsoft.CodeAnalysis.DocumentHighlighting
             return true;
         }
 
-        private static async Task AddLocationSpan(Location location, Solution solution, HashSet<DocumentSpan> spanSet, MultiDictionary<Document, HighlightSpan> tagList, HighlightSpanKind kind, CancellationToken cancellationToken)
+        private async Task AddLocationSpan(Location location, Solution solution, HashSet<DocumentSpan> spanSet, MultiDictionary<Document, HighlightSpan> tagList, HighlightSpanKind kind, CancellationToken cancellationToken)
         {
             var span = await GetLocationSpanAsync(solution, location, cancellationToken).ConfigureAwait(false);
             if (span != null && !spanSet.Contains(span.Value))
@@ -291,7 +262,7 @@ namespace Microsoft.CodeAnalysis.DocumentHighlighting
             }
         }
 
-        private static async Task<DocumentSpan?> GetLocationSpanAsync(
+        private async Task<DocumentSpan?> GetLocationSpanAsync(
             Solution solution, Location location, CancellationToken cancellationToken)
         {
             try
