+++ src/VisualStudio/IntegrationTest/TestUtilities/WellKnownTagNames.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
-using Microsoft.CodeAnalysis.Editor.ReferenceHighlighting;
+using Microsoft.CodeAnalysis.Editor.Implementation.ReferenceHighlighting;
 
 namespace Microsoft.VisualStudio.IntegrationTest.Utilities
 {
