+++ src/Workspaces/Remote/ServiceHub/ServiceHub.csproj
@@ -73,7 +73,6 @@
     </Compile>
     <Compile Include="Services\CodeAnalysisService_CodeLens.cs" />
     <Compile Include="Services\CodeAnalysisService_DesignerAttributes.cs" />
-    <Compile Include="Services\CodeAnalysisService_DocumentHighlights.cs" />
     <Compile Include="Services\CodeAnalysisService_TodoComments.cs" />
     <Compile Include="Services\CodeAnalysisService_SymbolFinder.cs" />
     <Compile Include="Services\RemoteSymbolSearchUpdateEngine.cs" />
